fish: initial work on `flake` function
this is a function that allows you to store a flake outside of the
git tree, making it much easier to use flakes in other projects that
don't plan on using nix.

The motivation for this is to make sure each project's environment
is totally isolated - I don't want to install tools for *anything*
globally. But that's hard to not do when adding a flake to a project
that doesn't use nix is so hacky. This is the best solution I've found.