Add license
I don't want to rewrite history by moving this to the very beginning,
but honestly it took me until now to even realize that I didn't have
a license in the repository. I always intended to make my configuration
"open source" in the ideology sense, not just the source-available way.

Can I just say that this license also retroactively applies to everything
that came before it? Probably not, but I'll believe it in my heart :)