Migrate Git config to imperative files managed by HM
Some configuration is easier to manage imperatively instead of
declaratively. Declarative means that any change to the configuration
requires a system rebuild, which is too heavyweight for a lot of things.

For example, I haven't tried out the Helix plugins branch because it
would require a lot of changes to the configuration, so using HM to
apply an imperatively managed configuration would allow me to iterate on
things in a faster feedback loop.

Some programs will still be managed by HM entirely, since there can be a
lot of advantages to doing so.