Refactor configuration to symlink directories
This commit fundamentally changes the way that the configuration is
defined. Instead of declaring everything using Nix, and having the
configuration compiled then applied (which requires a rebuild to change
anything on the system), config is instead symlinked to this repo.

This gives me full flexibility to change things without rebuilding,
and still gives me the control over how and where it is applied on each
system I control.

This commit also introduces the use of Blueprint, which removes all of
the glue code in my flake. As of the time of writing this (2025-01-14),
Blueprint does not support Home Manager configurations, but I am working
on adding support.