Make ghostty config better for pixel shift
I've been doing some of this work on my OLED TV, which means naturally
it has pixel shift to prevent burn-in. This means that with little
padding the text can get clipped at the edge of the TV.

Working with a 65" TV is fantastic. I plan to purchase one instead of a
monitor, and work exclusively from it instead of a traditional desk
setup.