Start building a tiny DDNS client
Scratching my own itch with this. I didn't need to build this myself,
but doing it in a compiled language felt better than using Python or
Perl.

CocoisBuggy/cloudflare_ddns is the best option available for CF users,
but after experimenting, I decided to roll my own because it's easier to
add more features in the long run if I'm not relying on someone else's
tiny project that only exists to scratch *their* itch.