Use macOS sshd
I was previously using Tailscale SSH to host the SSH server on macmini.
This was convenient but because of the web authentication it was mildly
frustrating every time I needed to SSH into this machine.

This commit remedies this by enabling the default SSH server and
configuring it to disallow password login. It authorizes macmini,
macbook-air, and my iPhone (via Blink). Eventually I'll add my iPad to
this list too, since I occasionally use it as a remote development
device.