home: refactor my.config.directory to take absolute paths
This allows me to use the path to the flake itself in the nix store as
the default, which means generic "users" can now use this config without
having to have a machine specific config. However, this will result in
immutable config files, since the links will be directly to the nix
store.