legacy: add configuration for my legacy PC
This is an older machine that I've installed a regular distro on but am
managing my user configuration with home manager. I've had a
configuration on this system for a while, but I haven't updated it since
before the big config rewrite. This is a much simpler diff to merge it
in.