nvim: init minimal configuration
This is a plugin management strategy that I haven't seen before, but
that's not to say that nobody is doing it. I'm using Lazy to initialize
the plugins, but they're downloaded and managed by my Home Manager
configuration. There is a file in my nvim directory that specifies the
versions and build directions for each package.

This gives me some nice benefits:

- On a new setup, there will never be a time before plugins are
  installed.
- Versions are locked, and upgrades are a very manual process.
  - This helps me get around the constant instability of updating
    plugins. Ugrading becomes intentional rather than something you do
    while you're waiting for something to compile, and accidentally
    breaking your configuration while you're at it.
- Builds are managed by Nix, instead of relying on your environment to
  have the right tools available.

There are also some obvious drawbacks, but the simplicity and ideology
of this management strategy outweighs it (for me, for now).