nvim: use built-in package system for loading, add module
This commit simplifies the plugin loading system by using the built-in
packages feature in Vim. Packages are a set of plugins, and these
plugins are either loaded automatically ('start') or optionally ('opt').

There is now a module (my.programs.neovim) which allows one to declare a
package (e.g., 'my.programs.neovim.packages.my-package-name') and a list
of plugins it supplies. In my config, this is used to declare a
'nix-managed' package, which contains all the plugins I use -- though
more packages could be added later, if necessary.

This also means that the configuration must now explicitly require()
each of the plugins. This conversion has not yet been completed, but is
in progress.