host: init work-macbookpro (hostname may change in the future)
This commit introduces a new host for my work-provided MacBook Pro, but
doesn't introduce anything new in the run script yet because I just
needed to get the setup working. Next time I need to apply it, I will
also adjust the script. It should be as simple as either duplicating or
abstracting the macmini function.