zsh: update zshrc
This commit skips executing fish if the terminal is Warp, because that
seems to break it. Every few months I open Warp again, remember why I
don't like it, then close it -- but during the time I have it open, it
doesn't work because of my goofy ass fish init system.